初步修改，更新文档

+++ infer.py
@@ -25,7 +25,7 @@ If you find no further external knowledge needed, you can directly provide the a
 
 # Initialize the tokenizer and model
 tokenizer = transformers.AutoTokenizer.from_pretrained(model_id)
-model = transformers.AutoModelForCausalLM.from_pretrained(model_id, torch_dtype=torch.bfloat16, device_map="auto")
+model = transformers.AutoModelForCausalLM.from_pretrained(model_id, torch_dtype=torch.float16, device_map="auto")
 
 # Define the custom stopping criterion
 class StopOnSequence(transformers.StoppingCriteria):
@@ -105,13 +105,14 @@ while True:
         do_sample=True,
         temperature=0.7
     )
-
+    #终止符中止
     if outputs[0][-1].item() in curr_eos:
         generated_tokens = outputs[0][input_ids.shape[1]:]
         output_text = tokenizer.decode(generated_tokens, skip_special_tokens=True)
         print(output_text)
         break
 
+    #提问终止，提取问题
     generated_tokens = outputs[0][input_ids.shape[1]:]
     output_text = tokenizer.decode(generated_tokens, skip_special_tokens=True)
